added more data to seed.

+++ database/migrations/2020_01_18_063918_create_meeting_times_table.php
@@ -18,7 +18,7 @@ class CreateMeetingTimesTable extends Migration
             $table->string('day');
             $table->string('start');
             $table->string('end');
-            $table->integer('duration');
+            $table->float('duration');
             $table->timestamps();
         });
     }
